Create a Translatable trait to reduce code duplication in entity classes
