+++ src/Pim/Bundle/CatalogBundle/Model/AbstractAttribute.php
@@ -9,7 +9,7 @@ use Pim\Bundle\CatalogBundle\Entity\AttributeGroup;
 use Pim\Bundle\CatalogBundle\Entity\Family;
 use Pim\Bundle\CatalogBundle\Entity\Locale;
 use Pim\Bundle\TranslationBundle\Entity\TranslatableInterface;
-use Pim\Bundle\TranslationBundle\Entity\AbstractTranslation;
+use Pim\Bundle\TranslationBundle\Entity\Translatable;
 use Pim\Bundle\VersioningBundle\Model\VersionableInterface;
 
 /**
@@ -22,6 +22,8 @@ use Pim\Bundle\VersioningBundle\Model\VersionableInterface;
 abstract class AbstractAttribute implements TimestampableInterface, TranslatableInterface,
  GroupSequenceProviderInterface, ReferableInterface, VersionableInterface
 {
+    use Translatable;
+
     /**
      * Attribute id
      * @var integer $id
@@ -171,17 +173,6 @@ abstract class AbstractAttribute implements TimestampableInterface, Translatable
 
     /** @var integer $minimumInputLength */
     protected $minimumInputLength = 0;
-
-    /**
-     * Used locale to override Translation listener's locale
-     * this is not a mapped field of entity metadata, just a simple property
-     *
-     * @var string $locale
-     */
-    protected $locale;
-
-    /** @var ArrayCollection $translations */
-    protected $translations;
 
     /**
      * Constructor
@@ -1235,70 +1226,6 @@ abstract class AbstractAttribute implements TimestampableInterface, Translatable
             }
             $this->$method($value);
         }
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setLocale($locale)
-    {
-        $this->locale = $locale;
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getTranslation($locale = null)
-    {
-        $locale = ($locale) ? $locale : $this->locale;
-        if (!$locale) {
-            return null;
-        }
-        foreach ($this->getTranslations() as $translation) {
-            if ($translation->getLocale() == $locale) {
-                return $translation;
-            }
-        }
-
-        $translationClass = $this->getTranslationFQCN();
-        $translation      = new $translationClass();
-        $translation->setLocale($locale);
-        $translation->setForeignKey($this);
-        $this->addTranslation($translation);
-
-        return $translation;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getTranslations()
-    {
-        return $this->translations;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function addTranslation(AbstractTranslation $translation)
-    {
-        if (!$this->translations->contains($translation)) {
-            $this->translations->add($translation);
-        }
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function removeTranslation(AbstractTranslation $translation)
-    {
-        $this->translations->removeElement($translation);
 
         return $this;
     }
